Test HTML reprs with cached JSON data #19
Rather than always testing against server data
pulled from Earth Engine, tests now run against
locally cached JSON objects if they exist. If
not, they are downloaded and added to the cache,
using the serialized object as the key to ensure
objects are unique in the cache. This means that
new tests can be written with new objects without
having to manually store new data.
